+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## vx.x.x - xxxx-xx-xx
 
+### Changes:
 
+* Added options for fleetcarrier cooldown notifications: Overlay, Popup, Both, None.
 
 
 ## v5.5.0-b1 - 2026-04-19
@@ -17,7 +19,6 @@
 * The API setup now has a checkbox to enable / disable the objectives API. Also, all the quick shortcut buttons have now been enhanced to set/unset this checkbox appropriately for each integration.
 * Removed the quick-setup button for the Spectrum integration as the service appears to be dead and the maintainer has vanished.
 * `/scripts` folder is now excluded from the distribution zip.
-* Added options for fleetcarrier cooldown notifications: Overlay, Popup, Both, None.
 
 ### Bug Fixes: