+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bug Fixes:
 
 * Show / hide system information overlay wasn't respecting the user's choice in settings.
+* Plugin failed to load on startup if fleet carrier data was incomplete.
 
 
 ## v5.5.0 - 2026-05-09