+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## vx.x.x - xxxx-xx-xx
+## v5.5.1 - 2026-05-11
 
 ### Bug Fixes: